requirements.txt
streamlit
openai
pandas
python-dotenv
geopy
streamlit-option-menu
pydeck
folium
streamlit-folium